+++ lib/regexec.c
@@ -179,7 +179,7 @@ tre_match(const tre_tnfa_t *tnfa, const void *string, size_t len,
 	       capabilities from the input stream. */
 	    return REG_BADPAT;
 	}
-      status = tre_tnfa_run_backtrack(tnfa, string, len, type,
+      status = tre_tnfa_run_backtrack(tnfa, string, (int)len, type,
 				      tags, eflags, &eo);
     }
 #ifdef TRE_APPROX
@@ -191,14 +191,14 @@ tre_match(const tre_tnfa_t *tnfa, const void *string, size_t len,
       regaparams_default(&params);
       params.max_err = 0;
       params.max_cost = 0;
-      status = tre_tnfa_run_approx(tnfa, string, len, type, tags,
+      status = tre_tnfa_run_approx(tnfa, string, (int)len, type, tags,
 				   &match, params, eflags, &eo);
     }
 #endif /* TRE_APPROX */
   else
     {
       /* Exact matching, no back references, use the parallel matcher. */
-      status = tre_tnfa_run_parallel(tnfa, string, len, type,
+      status = tre_tnfa_run_parallel(tnfa, string, (int)len, type,
 				     tags, eflags, &eo);
     }
 
@@ -226,7 +226,7 @@ int
 regexec(const regex_t *preg, const char *str,
 	size_t nmatch, regmatch_t pmatch[], int eflags)
 {
-  return regnexec(preg, str, -1, nmatch, pmatch, eflags);
+  return regnexec(preg, str, (unsigned)-1, nmatch, pmatch, eflags);
 }
 
 
@@ -244,7 +244,7 @@ int
 regwexec(const regex_t *preg, const wchar_t *str,
 	 size_t nmatch, regmatch_t pmatch[], int eflags)
 {
-  return regwnexec(preg, str, -1, nmatch, pmatch, eflags);
+  return regwnexec(preg, str, (unsigned)-1, nmatch, pmatch, eflags);
 }
 
 #endif /* TRE_WCHAR */
@@ -254,7 +254,7 @@ reguexec(const regex_t *preg, const tre_str_source *str,
 	 size_t nmatch, regmatch_t pmatch[], int eflags)
 {
   tre_tnfa_t *tnfa = (void *)preg->TRE_REGEX_T_FIELD;
-  return tre_match(tnfa, str, -1, STR_USER, nmatch, pmatch, eflags);
+  return tre_match(tnfa, str, (unsigned)-1, STR_USER, nmatch, pmatch, eflags);
 }
 
 
@@ -294,7 +294,7 @@ tre_match_approx(const tre_tnfa_t *tnfa, const void *string, size_t len,
       if (tags == NULL)
 	return REG_ESPACE;
     }
-  status = tre_tnfa_run_approx(tnfa, string, len, type, tags,
+  status = tre_tnfa_run_approx(tnfa, string, (int)len, type, tags,
 			       match, params, eflags, &eo);
   if (status == REG_OK)
     tre_fill_pmatch(match->nmatch, match->pmatch, tnfa->cflags, tnfa, tags, eo);
@@ -319,7 +319,7 @@ int
 regaexec(const regex_t *preg, const char *str,
 	 regamatch_t *match, regaparams_t params, int eflags)
 {
-  return reganexec(preg, str, -1, match, params, eflags);
+  return reganexec(preg, str, (unsigned)-1, match, params, eflags);
 }
 
 #ifdef TRE_WCHAR
@@ -337,7 +337,7 @@ int
 regawexec(const regex_t *preg, const wchar_t *str,
 	  regamatch_t *match, regaparams_t params, int eflags)
 {
-  return regawnexec(preg, str, -1, match, params, eflags);
+  return regawnexec(preg, str, (unsigned)-1, match, params, eflags);
 }
 
 #endif /* TRE_WCHAR */
